refactor: separate least_square from heavy
